+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+v0.1.3 (2025-04-15)
+-------------------
+
+- Use the value from the AltchaField `maxnumber` option, when provided, to generate the
+  challenge in `get_altcha_challenge`.
+  https://github.com/aboutcode-org/django-altcha/issues/5
+
 v0.1.2 (2025-03-31)
 -------------------